Refactor openapi support from Tool
